Split datastructures and made simulator part of testing framework
